Reformat code

+++ src/common/gameocean/inventory/Inventory.cpp
@@ -8,7 +8,7 @@ bool Inventory::addItem(ItemInstance item) {
         if (s.isSameType(item) && s.getItem()->getMaxStackSize() > s.count) {
             if (s.count + item.count > s.getItem()->getMaxStackSize()) {
                 item.count -= s.getItem()->getMaxStackSize() - s.count;
-                s.count = s.getItem()->getMaxStackSize();
+                s.count = (byte) s.getItem()->getMaxStackSize();
                 setItem(i, s);
             } else {
                 s.count += item.count;
@@ -23,7 +23,7 @@ bool Inventory::addItem(ItemInstance item) {
         if (s.isEmpty()) {
             s.setItem(item.getItem(), item.getItemData());
             if (item.count > s.getItem()->getMaxStackSize()) {
-                s.count = s.getItem()->getMaxStackSize();
+                s.count = (byte) s.getItem()->getMaxStackSize();
                 item.count -= s.count;
                 setItem(i, s);
             } else {
@@ -57,9 +57,9 @@ bool Inventory::removeItem(ItemInstance item) {
     return false;
 }
 
-bool Inventory::findItem(ItemInstance const& item, bool exact) {
+bool Inventory::findItem(ItemInstance const &item, bool exact) {
     int c = item.count;
-    for (ItemInstance& s : items) {
+    for (ItemInstance &s : items) {
         if (s.isSameType(item)) {
             if (exact) {
                 if (s.count == item.count)
